nano

+++ server_extensions_extra/src/main/java/com/generalbytes/batm/server/extensions/extra/nano/wallets/nano/BalanceData.java
@@ -19,7 +19,7 @@ public class BalanceData {
 
 
     public BigDecimal getBalance() {
-        BigInteger ten = BigInteger.TEN;
+        BigDecimal ten = BigDecimal.TEN;
         return balance.divide(ten.pow(8));
     }
     public String getConfirmations() {
